+++ CHANGES.rst
@@ -10,6 +10,8 @@ Unreleased
 -   The debugger escapes the exception message in the page title. :pr:`2719`
 -   When binding ``routing.Map``, a long IDNA ``server_name`` with a port does not fail
     encoding. :issue:`2700`
+-   ``iri_to_uri`` shows a deprecation warning instead of an error when passing bytes.
+    :issue:`2708`
 
 
 Version 2.3.4