+++ CHANGES.rst
@@ -3,7 +3,7 @@
 Version 2.3.5
 -------------
 
-Unreleased
+Released 2023-06-07
 
 -   Python 3.12 compatibility. :issue:`2704`
 -   Fix handling of invalid base64 values in ``Authorization.from_header``. :issue:`2717`